# lib/api-zod/src/generated/types/analysisSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisPoint } from './analysisPoint';
import type { Dashboard } from './dashboard';

export interface AnalysisSession {
  id: string;
  source: string;
  mode: string;
  status: string;
  progress: number;
  processedFrames: number;
  totalFrames: number;
  current: Dashboard;
  points: AnalysisPoint[];
  startedAt: string;
}