# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisPoint } from './analysisPoint';
import type { AnalyticsDistributionItem } from './analyticsDistributionItem';
import type { AnalyticsEvaluation } from './analyticsEvaluation';
import type { AnalyticsWeeklyAlertsItem } from './analyticsWeeklyAlertsItem';

export interface Analytics {
  points: AnalysisPoint[];
  distribution: AnalyticsDistributionItem[];
  weeklyAlerts: AnalyticsWeeklyAlertsItem[];
  evaluation: AnalyticsEvaluation;
}