# lib/api-zod/src/generated/types/cameraUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CameraUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  location?: string;
  source?: string;
  sourceType?: string;
  status?: string;
}