# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFactors } from './riskFactors';

export interface Dashboard {
  status: string;
  mode: string;
  currentCount: number;
  density: number;
  movementSpeed: number;
  direction: string;
  movementState: string;
  riskLevel: string;
  riskScore: number;
  activeAlerts: number;
  fps: number;
  source: string;
  factors: RiskFactors;
  updatedAt: string;
}