# lib/api-zod/src/generated/types/riskSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskSettingsWeights } from './riskSettingsWeights';

export interface RiskSettings {
  warningThreshold: number;
  highThreshold: number;
  criticalThreshold: number;
  weights: RiskSettingsWeights;
}