# lib/api-zod/src/generated/types/riskSettingsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskSettingsInputWeights } from './riskSettingsInputWeights';

export interface RiskSettingsInput {
  /**
     * @minimum 1
     * @maximum 99
     */
  warningThreshold: number;
  /**
     * @minimum 1
     * @maximum 99
     */
  highThreshold: number;
  /**
     * @minimum 1
     * @maximum 100
     */
  criticalThreshold: number;
  weights: RiskSettingsInputWeights;
}